Added tagging interfaces to generated API to enable generic processing of data components (+ other small mods in API)

+++ swe-common-core/src/main/java/net/opengis/swe/v20/BinaryBlock.java
@@ -99,7 +99,7 @@ public interface BinaryBlock extends AbstractSWE, BinaryMember
     /**
      * Gets the byteLength property
      */
-    public int getByteLength();
+    public long getByteLength();
     
     
     /**
@@ -111,7 +111,7 @@ public interface BinaryBlock extends AbstractSWE, BinaryMember
     /**
      * Sets the byteLength property
      */
-    public void setByteLength(int byteLength);
+    public void setByteLength(long byteLength);
     
     
     /**
